+++ stripe-einvoice-spring-boot-starter/src/main/java/com/housedevinci/einvoice/autoconfigure/EInvoiceDocumentAutoConfiguration.java
@@ -10,6 +10,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 
 /**
@@ -23,6 +24,7 @@ import org.springframework.context.annotation.Bean;
  * tried on.
  */
 @AutoConfiguration(before = EInvoiceIssuanceAutoConfiguration.class)
+@EnableConfigurationProperties(EInvoiceProperties.class)
 public class EInvoiceDocumentAutoConfiguration {
 
   private static final Logger log =